refactor: VotingCard
