+++ frontend/src/lib/components/proposal-detail/ProposalVotingSection.svelte
@@ -1,7 +1,6 @@
 <script lang="ts">
   import type { ProposalInfo } from "@dfinity/nns";
   import VotesResults from "./VotesResults.svelte";
-  import VotingCard from "./VotingCard/VotingCard.svelte";
   import { ProposalRewardStatus } from "@dfinity/nns";
   import { E8S_PER_ICP } from "$lib/constants/icp.constants";
   import { basisPointsToPercent } from "$lib/utils/utils";
@@ -9,6 +8,7 @@
     MINIMUM_YES_PROPORTION_OF_EXERCISED_VOTING_POWER,
     MINIMUM_YES_PROPORTION_OF_TOTAL_VOTING_POWER,
   } from "$lib/constants/proposals.constants";
+  import NnsVotingCard from "$lib/components/proposal-detail/VotingCard/NnsVotingCard.svelte";
 
   export let proposalInfo: ProposalInfo;
 
@@ -45,5 +45,5 @@
 />
 
 {#if !settled}
-  <VotingCard {proposalInfo} />
+  <NnsVotingCard {proposalInfo} />
 {/if}
