+++ frontend/src/lib/components/proposal-detail/VotingCard/VotingCard.svelte
@@ -1,131 +1,32 @@
 <script lang="ts">
-  import {
-    type ProposalInfo,
-    type Vote,
-    votableNeurons as getVotableNeurons,
-  } from "@dfinity/nns";
-
-  import { getContext } from "svelte";
-  import { definedNeuronsStore, neuronsStore } from "$lib/stores/neurons.store";
   import VotingConfirmationToolbar from "./VotingConfirmationToolbar.svelte";
-  import VotingNeuronSelect from "./VotingNeuronSelect.svelte";
-  import {
-    SELECTED_PROPOSAL_CONTEXT_KEY,
-    type SelectedProposalContext,
-  } from "$lib/types/selected-proposal.context";
-  import {
-    isProposalDeadlineInTheFuture,
-    nnsNeuronToVotingNeuron,
-  } from "$lib/utils/proposals.utils";
-  import {
-    voteRegistrationStore,
-    type VoteRegistrationStoreEntry,
-  } from "$lib/stores/vote-registration.store";
+  import { type VoteRegistrationStoreEntry } from "$lib/stores/vote-registration.store";
   import { BottomSheet } from "@dfinity/gix-components";
   import { i18n } from "$lib/stores/i18n";
   import SignInGuard from "$lib/components/common/SignInGuard.svelte";
   import SpinnerText from "$lib/components/ui/SpinnerText.svelte";
-  import { OWN_CANISTER_ID_TEXT } from "$lib/constants/canister-ids.constants";
-  import { votingNeuronSelectStore } from "$lib/stores/vote-registration.store";
-  import { registerNnsVotes } from "$lib/services/nns-vote-registration.services";
   import MyVotes from "$lib/components/proposal-detail/MyVotes.svelte";
   import IneligibleNeuronsCard from "$lib/components/proposal-detail/IneligibleNeuronsCard.svelte";
   import VotingNeuronSelectList from "$lib/components/proposal-detail/VotingCard/VotingNeuronSelectList.svelte";
   import {
     type CompactNeuronInfo,
-    filterIneligibleNnsNeurons,
     type IneligibleNeuronData,
     neuronsVotingPower,
-    votedNeuronDetails,
   } from "$lib/utils/neuron.utils";
   import { NNS_MINIMUM_DISSOLVE_DELAY_TO_VOTE } from "$lib/constants/neurons.constants";
   import { authSignedInStore } from "$lib/derived/auth.derived";
-  import { isForceCallStrategy } from "$lib/utils/env.utils";
+  import VotingNeuronSelect from "$lib/components/proposal-detail/VotingCard/VotingNeuronSelect.svelte";
 
-  export let proposalInfo: ProposalInfo;
+  export let hasNeurons: boolean;
+  export let visible: boolean;
+  export let neuronsReady: boolean;
+  export let voteRegistration: VoteRegistrationStoreEntry | undefined;
+  export let neuronsVotedForProposal: CompactNeuronInfo[];
+  export let ineligibleNeurons: IneligibleNeuronData[];
+  export let minSnsDissolveDelaySeconds: bigint;
 
-  const votableNeurons = () =>
-    getVotableNeurons({
-      neurons: $definedNeuronsStore,
-      proposal: proposalInfo,
-    }).map((neuron) =>
-      nnsNeuronToVotingNeuron({ neuron, proposal: proposalInfo })
-    );
-
-  let visible = false;
-  /** Signals that the initial checkbox preselection was done. To avoid removing of user selection after second queryAndUpdate callback. */
-  let initialSelectionDone = false;
-  let voteRegistration: VoteRegistrationStoreEntry | undefined = undefined;
-
-  $: voteRegistration = (
-    $voteRegistrationStore.registrations[OWN_CANISTER_ID_TEXT] ?? []
-  ).find(({ proposalIdString }) => `${proposalInfo.id}` === proposalIdString);
-
-  $: $definedNeuronsStore,
-    (visible = isProposalDeadlineInTheFuture(proposalInfo));
-
-  const updateVotingNeuronSelectedStore = () => {
-    if (!initialSelectionDone) {
-      initialSelectionDone = true;
-      votingNeuronSelectStore.set(votableNeurons());
-    } else {
-      // preserve user selection after neurons update (e.g. queryAndUpdate second callback)
-      votingNeuronSelectStore.updateNeurons(votableNeurons());
-    }
-  };
-
-  $: $definedNeuronsStore, updateVotingNeuronSelectedStore();
-
-  const { store } = getContext<SelectedProposalContext>(
-    SELECTED_PROPOSAL_CONTEXT_KEY
-  );
-  const vote = async ({ detail }: { detail: { voteType: Vote } }) =>
-    await registerNnsVotes({
-      neuronIds: $votingNeuronSelectStore.selectedIds.map(BigInt),
-      vote: detail.voteType,
-      proposalInfo,
-      reloadProposalCallback: (
-        proposalInfo: ProposalInfo // we update store only if proposal id are matching even though it would be an edge case that these would not match here
-      ) =>
-        store.update(({ proposalId, proposal }) => ({
-          proposalId,
-          proposal: proposalId === proposalInfo.id ? proposalInfo : proposal,
-        })),
-    });
-
-  // UI loader
-  const neuronsStoreReady = (): boolean => {
-    // We consider the neurons store as ready if it has been initialized once. Subsequent changes that happen after vote or other functions are handled with the busy store.
-    // This to avoid the display of a spinner within the page and another spinner over it (the busy spinner) when the user vote is being processed.
-    if (neuronsReady) {
-      return true;
-    }
-
-    return (
-      $neuronsStore.neurons !== undefined &&
-      ($neuronsStore.certified === true ||
-        ($neuronsStore.certified === false && isForceCallStrategy()))
-    );
-  };
-
-  let neuronsReady = false;
-  $: $neuronsStore, (neuronsReady = neuronsStoreReady());
-
-  let neuronsVotedForProposal: CompactNeuronInfo[];
-  $: {
-    neuronsVotedForProposal = votedNeuronDetails({
-      neurons: $definedNeuronsStore,
-      proposal: proposalInfo,
-    });
-  }
   let votedVotingPower: bigint;
   $: votedVotingPower = neuronsVotingPower(neuronsVotedForProposal);
-
-  let ineligibleNeurons: IneligibleNeuronData[];
-  $: ineligibleNeurons = filterIneligibleNnsNeurons({
-    neurons: $definedNeuronsStore,
-    proposal: proposalInfo,
-  });
 </script>
 
 <BottomSheet>
@@ -135,13 +36,10 @@
     data-tid="voting-card-component"
   >
     <SignInGuard>
-      {#if $definedNeuronsStore.length > 0}
+      {#if hasNeurons}
         {#if neuronsReady}
           {#if visible}
-            <VotingConfirmationToolbar
-              {voteRegistration}
-              on:nnsConfirm={vote}
-            />
+            <VotingConfirmationToolbar {voteRegistration} on:nnsConfirm />
           {/if}
 
           <VotingNeuronSelect
@@ -154,9 +52,7 @@
             <IneligibleNeuronsCard
               slot="ineligible-neurons"
               {ineligibleNeurons}
-              minSnsDissolveDelaySeconds={BigInt(
-                NNS_MINIMUM_DISSOLVE_DELAY_TO_VOTE
-              )}
+              {minSnsDissolveDelaySeconds}
             />
           </VotingNeuronSelect>
         {:else}
@@ -173,14 +69,12 @@
 <style lang="scss">
   @use "@dfinity/gix-components/dist/styles/mixins/media";
 
-  // TODO(max): combine w/ sns version?
   .container {
     display: flex;
 
     // add scrollbars for too long content
-    max-height: 100vh;
     overflow-y: auto;
-
+    max-height: 100vh;
     @include media.min-width(large) {
       max-height: none;
     }
